requirements.txt
addict==2.4.0
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
brotlicffi==1.2.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==4.0.0
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.3
Flask==3.1.2
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
gmpy2==2.2.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
ipython==8.38.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mkl_fft==2.1.1
mkl_random==1.3.0
mkl-service==2.5.2
mpmath==1.3.0
narwhals==2.16.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
open3d==0.19.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
pip==25.3
platformdirs==4.5.1
plotly==6.5.2
polyscope==2.5.0
prompt_toolkit==3.0.52
protobuf==6.33.5
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rpds-py==0.30.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.51.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.3
traitlets==5.14.3
trimesh==4.11.1
triton==3.1.0
typer-slim==0.21.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
wandb==0.24.1
wcwidth==0.5.3
Werkzeug==3.1.5
wheel==0.46.3
widgetsnbextension==4.0.15
zipp==3.23.0
